starting to refactor
